pyproject.toml
[project]
name = "ozonlayer-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alexeev Bronislav", email = "alexeev.dev@mail.ru" }
]
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "orjson>=3.11.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "selenium>=4.39.0",
    "undetected-chromedriver>=3.5.5",
]

[project.scripts]
ozonlayer-scraper = "ozonlayer_scraper:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
